fix(InstallFcitx5):

+++ InstallFcitx5.sh
@@ -1,4 +1,18 @@
-yay -S --needed fcitx5 fcitx5qt fcitx5gtk fcitx5-configtool fcitx5-chinese-addons fcitx5-pinyin-zhwiki
+if [[ "$EUID" != 0 ]]; then
+	echo "Run this in sudo"
+	exit 1
+fi
+read -p "Input your username if want to install fcitx5:" uname
+if [[ -z $uname ]]; then
+	echo "Aborted"
+	exit
+fi
+sudo -u $uname yay -S --needed fcitx5 fcitx-5qt fcitx5-gtk fcitx5-configtool fcitx5-chinese-addons fcitx5-pinyin-zhwiki
+read -p "Export environment variables?(y/N)" ienv
+if [[ ienv != [yY] ]]; then
+	echo "Exit"
+	exit
+fi
 tee -a >>/etc/environment <<EOF
 #fcitx
 #GTK_IM_MODULE=fcitx
